fix(security): harden three leftover findings in ssap/hadm/sle utils
- ssaps_server_find: guard mtu lower bound before leftSize subtraction in
  BuildMixPrimaryServicePayloadV10 (mirror the non-V10 guard), preventing
  uint32 underflow from driving a bogus realSize
- hadm client: compute maxLatency * minInterval in uint32 to avoid signed
  int overflow UB with peer-controlled connection parameters
- sle_service_data: read DIS AD length as an unsigned byte so the existing
  bounds check holds on signed-char platforms and the parse cursor can no
  longer wrap back into an infinite loop

+++ utils/src/sle_service_data.cpp
@@ -1148,7 +1148,7 @@ void SlePeripheralDevice::ParseSleDisData()
     size_t length = 0;
     while ((i + SLE_ADV_DEVICE_INFO_MIN_LEN) < disServiceData.size()) {
         type = disServiceData[i++];
-        length = disServiceData[i++];
+        length = static_cast<uint8_t>(disServiceData[i++]);
         if (length <= 0 || i + length > disServiceData.size()) {
             break;
         }
